requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-core
pdfminer.six
tiktoken
google-generativeai
python-dotenv
opensearch-py